fix: use float16 on CPU to prevent OOM segfault with TADA 3B
TADA 3B (4.2B params) needs ~16GB in float32, causing segfault on
32GB systems. float16 halves it to ~8GB and loads in 23s.

+++ backend.py
@@ -701,7 +701,8 @@ def load_model(model_id=None):
 
     cfg = MODELS.get(model_id, MODELS["tada-1b"])
     device = _get_device()
-    dtype = torch.bfloat16 if device == "cuda" else torch.float32
+    # Use float16 on CPU (float32 OOMs on 3B model), bfloat16 on CUDA
+    dtype = torch.bfloat16 if device == "cuda" else torch.float16
 
     with tada_lock:
         if tada_model is None:
